third_party/blink/web_tests/http/tests/wasm_streaming/wasm_navigate_while_streaming_compiling.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>

promise_test(async test => {

  const url1 = 'resources/wasm_load.html?small.wasm';
  const url2 = 'resources/wasm_load.html?large.wasm';

  const popup = window.open('about:blank');
  test.add_cleanup(() => popup.close());

  let navigation_resolve;
  window.addEventListener('message', (event) => {
    console.log(`[parent] got message: ${event.data}`);
    assert_equals(event.data.substring(0, 18), 'started compiling ');
    navigation_resolve();
  });

  const timeout = ms => new Promise(resolve => setTimeout(resolve, ms));

  for (let i = 0; i < 20; ++i) {
    let navigationPromiseResolve;
    const navigationPromise = new Promise((resolve) => {navigation_resolve = resolve});
    const url = i % 2 ? url1 : url2;
    console.log(`[parent] Iteration ${i}, navigating popup to ${url}`);
    popup.location.href = url;
    await navigationPromise;
    const ms = Math.floor(Math.random() * 50);
    console.log(`[parent] Waiting for ${ms}ms`);
    if (ms) await timeout(ms);
  }

  popup.close();
}, "Test navigating away while Wasm streaming compilation is running");

</script>
